+++ src/main/java/com/team2/resumeeditorproject/admin/controller/LandingController.java
@@ -16,8 +16,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createBadRequestResponse;
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createOkResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createBadRequestResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createOkResponse;
 
 @Controller
 @RequestMapping("/landing")
