+++ src/main/java/com/team2/resumeeditorproject/admin/controller/ResumeManagementController.java
@@ -23,8 +23,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createOkResponse;
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createPagedResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createOkResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createPagedResponse;
 
 @Controller
 @RequestMapping("/admin/resume")
