+++ src/main/java/com/team2/resumeeditorproject/admin/controller/ResumeStatisticsController.java
@@ -13,8 +13,8 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.function.Function;
 
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createBadRequestResponse;
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createOkResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createBadRequestResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createOkResponse;
 
 @Controller
 @RequestMapping("/admin/statistics") // /admin 지우기
