+++ src/main/java/com/team2/resumeeditorproject/admin/controller/ReviewManagementController.java
@@ -19,7 +19,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createPagedResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createPagedResponse;
 
 @Controller
 @RequestMapping("/admin/review")
