+++ src/main/java/com/team2/resumeeditorproject/user/controller/MailController.java
@@ -21,8 +21,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createBadRequestResponse;
-import static com.team2.resumeeditorproject.admin.service.ResponseHandler.createOkResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createBadRequestResponse;
+import static com.team2.resumeeditorproject.common.util.ResponseHandler.createOkResponse;
 
 @Controller
 @RequiredArgsConstructor
